# packages/api-contract/openapi.json
{
  "components": {
    "schemas": {
...
      "DomainIn": {
        "properties": {
          "host": {
            "maxLength": 253,
            "minLength": 4,
            "title": "Host",
            "type": "string"
          }
        },
        "required": [
          "host"
        ],
        "title": "DomainIn",
        "type": "object"
      },
      "DomainOut": {
        "properties": {
          "host": {
            "title": "Host",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "is_primary": {
            "title": "Is Primary",
            "type": "boolean"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "host",
          "kind",
          "status",
          "is_primary"
        ],
        "title": "DomainOut",
        "type": "object"
      },
      "DomainWithInstructions": {
        "properties": {
          "dns": {
            "additionalProperties": true,
            "title": "Dns",
            "type": "object"
          },
          "host": {
            "title": "Host",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "is_primary": {
            "title": "Is Primary",
            "type": "boolean"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "host",
          "kind",
          "status",
          "is_primary",
          "dns"
        ],
        "title": "DomainWithInstructions",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "ecomm SaaS marketplace API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/admin/domains": {
      "get": {
        "operationId": "list_domains_api_v1_admin_domains_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DomainOut"
                  },
                  "title": "Response List Domains Api V1 Admin Domains Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Domains",
        "tags": [
          "admin:domains"
        ]
      },
      "post": {
        "operationId": "add_domain_api_v1_admin_domains_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DomainIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DomainWithInstructions"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Domain",
        "tags": [
          "admin:domains"
        ]
      }
    },
    "/api/v1/admin/domains/{domain_id}/primary": {
      "post": {
        "operationId": "make_primary_api_v1_admin_domains__domain_id__primary_post",
        "parameters": [
          {
            "in": "path",
            "name": "domain_id",
            "required": true,
            "schema": {
              "title": "Domain Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DomainOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Make Primary",
        "tags": [
          "admin:domains"
        ]
      }
    },
    "/api/v1/admin/domains/{domain_id}/verify": {
      "post": {
        "operationId": "verify_domain_api_v1_admin_domains__domain_id__verify_post",
        "parameters": [
          {
            "in": "path",
            "name": "domain_id",
            "required": true,
            "schema": {
              "title": "Domain Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DomainOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Verify Domain",
        "tags": [
          "admin:domains"
        ]
      }
    },
...
  }
}